Removing deprecated tests
